refactor: Controller 제거 및 LottoGame으로 개명
- LottoGameController → LottoGame으로 클래스명 변경
- controller 패키지 제거 (불필요한 계층)
- Application 단순화 (의존성 조립 코드 제거)
- run() → start() 메서드명 변경
- LottoGame에 기본 생성자 추가 (의존성 내부 생성)
- 테스트를 위한 생성자 주입 생성자 유지

개선 효과:
- 웹 MVC 용어(Controller)를 도메인 용어(Game)로 변경
- 불필요한 패키지 계층 1개 제거
- Application의 역할이 명확해짐 (진입점)
- 작은 콘솔 앱에 적합한 구조로 개선

설계 철학:
"Controller는 웹 MVC의 개념이다. 단순한 콘솔 앱에서는
도메인 용어가 더 적합하며, 클래스 1개를 위한 패키지는 불필요하다."
